requirements.txt
streamlit
openai
faiss-cpu
PyMuPDF
langchain